refactor(mcp): dedup monitor research developer and test helpers
