+++ tests/mcp-smoke.test.mjs
@@ -1,53 +1,15 @@
 import assert from 'node:assert/strict';
-import { spawn } from 'node:child_process';
 import { createServer } from 'node:http';
-import net from 'node:net';
 import test from 'node:test';
 import { setTimeout as delay } from 'node:timers/promises';
 import { assertAgentMetadataPolicy } from '../scripts/agent-metadata-policy.mjs';
-
-async function getFreePort() {
-  const server = net.createServer();
-  await new Promise((resolve, reject) => {
-    server.once('error', reject);
-    server.listen(0, '127.0.0.1', resolve);
-  });
-  const address = server.address();
-  assert.equal(typeof address, 'object');
-  const port = address.port;
-  await new Promise((resolve, reject) => {
-    server.close((error) => (error ? reject(error) : resolve()));
-  });
-  return port;
-}
-
-async function waitForHealth(port, child) {
-  const url = `http://127.0.0.1:${port}/health`;
-  let lastError;
-  for (let i = 0; i < 60; i += 1) {
-    if (child.exitCode !== null) {
-      throw new Error(`server exited early with code ${child.exitCode}`);
-    }
-    try {
-      const response = await fetch(url);
-      if (response.ok) return response;
-      lastError = new Error(`health returned ${response.status}`);
-    } catch (error) {
-      lastError = error;
-    }
-    await delay(100);
-  }
-  throw lastError ?? new Error('server did not become healthy');
-}
-
-
-function parseSseJson(body) {
-  const dataLine = body
-    .split(/\r?\n/)
-    .find((line) => line.startsWith('data: '));
-  assert.ok(dataLine, `Missing SSE data line in body: ${body}`);
-  return JSON.parse(dataLine.slice('data: '.length));
-}
+import {
+  getFreePort,
+  parseSseJson,
+  spawnServer,
+  stopChild,
+  waitForHealth,
+} from './helpers.mjs';
 
 function assertServerGeneratedRequestId(payload, untrustedValues = []) {
   assert.match(
@@ -99,32 +61,6 @@ function assertKeylessAccountRecovery(
     result.structuredContent,
     untrustedRequestIds
   );
-}
-
-function spawnServer(env) {
-  const child = spawn(process.execPath, ['dist/index.js'], {
-    env: {
-      ...process.env,
-      MCP_DELEGATED_CREDENTIAL_SECRET:
-        'test-mcp-delegated-credential-secret-32',
-      ...env,
-    },
-    stdio: ['pipe', 'pipe', 'pipe'],
-  });
-  child.stderr.setEncoding('utf8');
-  child.stdout.setEncoding('utf8');
-  return child;
-}
-
-async function stopChild(child) {
-  if (child.exitCode !== null) return;
-  child.kill('SIGTERM');
-  await Promise.race([
-    new Promise((resolve) => child.once('exit', resolve)),
-    delay(2_000).then(() => {
-      if (child.exitCode === null) child.kill('SIGKILL');
-    }),
-  ]);
 }
 
 async function startFakeFirecrawlApi() {
